refactor: move several packages to internal folder
