+++ internal/ascii/block.go
