+++ internal/ascii/block_test.go
