+++ internal/ascii/text.go
