+++ internal/ascii/text_test.go
