+++ internal/github/client_test.go
@@ -4,9 +4,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/github/gh-skyline/errors"
-	"github.com/github/gh-skyline/testutil/mocks"
-	"github.com/github/gh-skyline/types"
+	"github.com/github/gh-skyline/internal/errors"
+	"github.com/github/gh-skyline/internal/testutil/mocks"
+	"github.com/github/gh-skyline/internal/types"
 )
 
 func TestGetAuthenticatedUser(t *testing.T) {
