+++ internal/stl/geometry/assets/invertocat.png
