+++ internal/stl/geometry/assets/monasans-medium.ttf
