+++ internal/stl/geometry/assets/monasans-regular.ttf
