+++ internal/stl/geometry/assets_test.go
