+++ internal/stl/geometry/text_test.go
